BACKEND WORK IN PROGRESS

+++ database/migrations/2025_06_01_101642_add_columns_to_tbl_users.php
@@ -9,15 +9,24 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('tbl_users', function (Blueprint $table) {
-            $table->string('avatar')->after('id');
-            $table->string('username')->after('avatar');
+            if (!Schema::hasColumn('tbl_users', 'avatar')) {
+                $table->string('avatar')->nullable()->after('id');
+            }
+            if (!Schema::hasColumn('tbl_users', 'username')) {
+                $table->string('username')->nullable()->after('avatar');
+            }
         });
     }
 
     public function down(): void
     {
         Schema::table('tbl_users', function (Blueprint $table) {
-            $table->dropColumn('username');
+            if (Schema::hasColumn('tbl_users', 'avatar')) {
+                $table->dropColumn('avatar');
+            }
+            if (Schema::hasColumn('tbl_users', 'username')) {
+                $table->dropColumn('username');
+            }
         });
     }
 };
